Reference Issues/PRs None
What does this implementation fix/change?
We are using the aggregate to group associated terms together, so they can be referenced by other functions using the root. They KeyWord search terms for the twitter scraper to use are now grouped and referenced by the KeywordSearch class. This is mostly to clean up the class and makes it easier to read and debug.

Any other comments?
No