Update BingCovid.py
This is an exercise of the AGGREGATES pattern. This file is a good contact material. It can be seen that there is a lot of data here. This is in line with the database problem mentioned in the lecture. In order to avoid a large amount of data stacking, it is necessary to aggregate here. And it can be seen that in the original project, the data that wants to use the Bing website needs to be referenced through another file. If it can be directly compressed into a list here, it will save more time.